Stop running tests in parallel
The tests are so quick that we spend more time in setting up the workers
than in running the tests. Also, the untarring of the test repo
to a fixed location causes race conditions and fixing it to untar
into a temp directory in each worker and then clean up seems
overkill.